Remove unnecessary error variant `InvalidHashFunction`
Since `StructOpt` automatically checks that the input string for 
`function` is among the valid options before parsing it, we don't need a 
seperate error for the case where `from_str` encounters an invalid 
string. If this occurs, it is a bug. Thus, simply panic in this case.

Additionally, update tests for `HashFunction::from_str` to test new 
behaviour.