Add `unwrap_or_exit()` function to HshResult for easier error handling
Add `unwrap_or_exit()` function to HshResult type for easier and cleaner 
error handling, along with `exitcode()` function of HshErr.

Ideally, `unwrap_or_exit()` would be implemented directly on HshResult, 
but, in order to get around the orphan rules, it must be part of a 
locally-defined trait implemented on HshResult.